Merged Window and Application.
I made Window creation come from Application so a window can have a reference to the application in the future. It will hopefully get rid of the singleton on Linux and possibly allow for menubars to be moved from Application to Window. I also fixed window close handlers not being called when Window::close is used on MacOS.